refactor(core): move TT UI elements to own directory
