+++ core/src/apps/binance/layout.py
@@ -8,8 +8,8 @@ from trezor.messages import (
     ButtonRequestType,
 )
 from trezor.strings import format_amount
-from trezor.ui.scroll import Paginated
-from trezor.ui.text import Text
+from trezor.ui.model.tt.scroll import Paginated
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_hold_to_confirm
 from apps.common.layout import split_address
