+++ core/src/apps/bitcoin/authorize_coinjoin.py
@@ -4,7 +4,7 @@ from trezor import ui
 from trezor.messages.AuthorizeCoinJoin import AuthorizeCoinJoin
 from trezor.messages.Success import Success
 from trezor.strings import format_amount
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.base import set_authorization
 from apps.common.confirm import require_confirm, require_hold_to_confirm
