+++ core/src/apps/common/signverify.py
@@ -2,7 +2,7 @@ from ubinascii import hexlify
 
 from trezor import utils, wire
 from trezor.crypto.hashlib import blake256, sha256
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
 from apps.common.layout import split_address
