+++ core/src/apps/ethereum/layout.py
@@ -3,7 +3,7 @@ from ubinascii import hexlify
 from trezor import ui
 from trezor.messages import ButtonRequestType
 from trezor.strings import format_amount
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 from trezor.utils import chunks
 
 from apps.common.confirm import require_confirm, require_hold_to_confirm
