+++ core/src/apps/management/change_wipe_code.py
@@ -2,8 +2,8 @@ from storage.device import is_initialized
 from trezor import config, ui, wire
 from trezor.messages.Success import Success
 from trezor.pin import pin_to_int
+from trezor.ui.model.tt.text import Text
 from trezor.ui.popup import Popup
-from trezor.ui.text import Text
 
 from apps.common.confirm import require_confirm
 from apps.common.layout import show_success
