+++ core/src/apps/management/get_next_u2f_counter.py
@@ -3,7 +3,7 @@ from trezor import ui, wire
 from trezor.messages import ButtonRequestType
 from trezor.messages.GetNextU2FCounter import GetNextU2FCounter
 from trezor.messages.NextU2FCounter import NextU2FCounter
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
