+++ core/src/apps/management/sd_protect.py
@@ -5,7 +5,7 @@ from trezor.crypto import random
 from trezor.messages import SdProtectOperationType
 from trezor.messages.Success import Success
 from trezor.pin import pin_to_int
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
 from apps.common.layout import show_success
