+++ core/src/apps/management/set_u2f_counter.py
@@ -3,7 +3,7 @@ from trezor import ui, wire
 from trezor.messages import ButtonRequestType
 from trezor.messages.SetU2FCounter import SetU2FCounter
 from trezor.messages.Success import Success
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
