+++ core/src/apps/misc/get_entropy.py
@@ -1,7 +1,7 @@
 from trezor.crypto import random
 from trezor.messages import ButtonRequestType
 from trezor.messages.Entropy import Entropy
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
