+++ core/src/apps/misc/sign_identity.py
@@ -3,7 +3,7 @@ from ustruct import pack, unpack
 from trezor import ui
 from trezor.crypto.hashlib import sha256
 from trezor.messages.SignedIdentity import SignedIdentity
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 from trezor.utils import chunks
 
 from apps.common import HARDENED, coins
