+++ core/src/apps/stellar/operations/layout.py
@@ -13,7 +13,7 @@ from trezor.messages.StellarManageOfferOp import StellarManageOfferOp
 from trezor.messages.StellarPathPaymentOp import StellarPathPaymentOp
 from trezor.messages.StellarPaymentOp import StellarPaymentOp
 from trezor.messages.StellarSetOptionsOp import StellarSetOptionsOp
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 from trezor.wire import ProcessError
 
 from apps.stellar import consts, helpers
