+++ core/src/apps/webauthn/add_resident_credential.py
@@ -2,7 +2,7 @@ import storage.device
 from trezor import ui, wire
 from trezor.messages.Success import Success
 from trezor.messages.WebAuthnAddResidentCredential import WebAuthnAddResidentCredential
-from trezor.ui.text import Text
+from trezor.ui.model.tt.text import Text
 
 from apps.common.confirm import require_confirm
 from apps.webauthn.confirm import ConfirmContent, ConfirmInfo
