--- core/src/trezor/ui/button.py
@@ -1,2 +1,0 @@
-# FIXME: temporarily reexport from new location for compatibility
-from .model.tt.button import *  # noqa: F401, F403
