--- core/src/trezor/ui/confirm.py
@@ -1,2 +1,0 @@
-# FIXME: temporarily reexport from new location for compatibility
-from .model.tt.confirm import *  # noqa: F401, F403
