+++ core/src/trezor/ui/model/tt/info.py
@@ -1,11 +1,12 @@
 from trezor import res, ui
-from trezor.ui.button import Button, ButtonConfirm
-from trezor.ui.confirm import CONFIRMED
-from trezor.ui.text import TEXT_LINE_HEIGHT, TEXT_MARGIN_LEFT, render_text
+
+from .button import Button, ButtonConfirm
+from .confirm import CONFIRMED
+from .text import TEXT_LINE_HEIGHT, TEXT_MARGIN_LEFT, render_text
 
 if False:
     from typing import Type
-    from trezor.ui.button import ButtonContent
+    from .button import ButtonContent
 
 
 class DefaultInfoConfirm:
