+++ core/src/trezor/ui/model/tt/swipe.py
