--- core/src/trezor/ui/text.py
@@ -1,2 +1,0 @@
-# FIXME: temporarily reexport from new location for compatibility
-from .model.tt.text import *  # noqa: F401, F403
